Do not rename choice types on StructureDefinitions (#972)
* Do not rename choice types on StructureDefinitions

A specific choice type on a StructureDefinition should not have its id
or path renamed. Instead, put a slice name on the choice element.
Specific choice types on Instances are not changed.

* Remove extra id assignment

* Move shortcut id function to where it is used

The shortcut id is only ever needed by setImpliedPropertiesOnInstance,
so the function to get the shortcut id is moved to be in the same file.
The constants used by this function are still in the ElementDefinition
file, since I feel like that is where I would expect to find them if I
were looking for them.

* Move choice type slice name postfixes

* Always include type on diff for choice element slice

The IG publisher requires that the differential element representing a
slice of a choice element always contain a type. This is the case even
if its type is the same as the type of the choice element. Add a check
when calculating the differential so that the type is always included
for these elements.

Added TODO note to ElementDefinitionType.fromJSON regarding undefined
properties. There may be upcoming changes to requirements regarding
these types, which are currently included due to the implementation
difference between how ElementDefinitionType's constructor and fromJSON
methods handle undefined properties.
