+++ test/fhirtypes/ElementDefinition.test.ts
@@ -7,6 +7,7 @@ import { Type } from '../../src/utils/Fishable';
 import { Invariant, FshCode } from '../../src/fshtypes';
 import path from 'path';
 import { cloneDeep } from 'lodash';
+import { OnlyRule } from '../../src/fshtypes/rules';
 
 describe('ElementDefinition', () => {
   let defs: FHIRDefinitions;
@@ -383,47 +384,6 @@ describe('ElementDefinition', () => {
       expect(Object.keys(diff.toJSON())).toHaveLength(3);
     });
 
-    it('should calculate diff id using shortcut syntax for a choice slice', () => {
-      valueX.sliceIt('type', '$this', false, 'open');
-      const valueString = valueX.addSlice('valueString', new ElementDefinitionType('string'));
-      const diff = valueString.calculateDiff();
-      // snapshot should retain formal syntax and slicename
-      expect(valueString.id).toBe('Observation.value[x]:valueString');
-      expect(valueString.path).toBe('Observation.value[x]');
-      expect(valueString.sliceName).toBe('valueString');
-      // differential should use shortcut syntax and remove slicename
-      expect(diff.id).toBe('Observation.valueString');
-      expect(diff.path).toBe('Observation.valueString');
-      expect(diff.sliceName).toBeUndefined();
-    });
-
-    it('should calculate diff id using shortcut syntax when "value[x]" is in the middle of the path', () => {
-      const valueStringElement = new ElementDefinition('Observation.value[x]:valueString.id');
-      const diff = valueStringElement.calculateDiff();
-      // snapshot should retain formal syntax
-      expect(valueStringElement.id).toBe('Observation.value[x]:valueString.id');
-      expect(valueStringElement.path).toBe('Observation.value[x].id');
-
-      // differential should use shortcut syntax
-      expect(diff.id).toBe('Observation.valueString.id');
-      expect(diff.path).toBe('Observation.valueString.id');
-    });
-
-    it('should not calculate diff id using shortcut syntax for a non-choice slices on value[x] elements', () => {
-      valueX.sliceIt('type', '$this', false, 'open');
-      const testSlice = valueX.addSlice('testSlice', new ElementDefinitionType('reference'));
-      const diff = testSlice.calculateDiff();
-      // snapshot should retain formal syntax and slicename
-      expect(testSlice.id).toBe('Observation.value[x]:testSlice');
-      expect(testSlice.path).toBe('Observation.value[x]');
-      expect(testSlice.sliceName).toBe('testSlice');
-
-      // differential not should use shortcut syntax
-      expect(diff.id).toBe('Observation.value[x]:testSlice');
-      expect(diff.path).toBe('Observation.value[x]');
-      expect(diff.sliceName).toBe('testSlice');
-    });
-
     it('should include only new constraints in a diff when constraints are added', () => {
       const myInvariant = new Invariant('inv-1');
       myInvariant.severity = new FshCode('warning');
@@ -471,6 +431,20 @@ describe('ElementDefinition', () => {
       );
       const diff = valueX.calculateDiff();
       expect(diff.mapping).toBeUndefined();
+    });
+
+    it('should always include type in the diff for a slice of a choice element', () => {
+      // constrain value[x] to Quantity so that, when sliced, the original on the slice will have only that type
+      const onlyRule = new OnlyRule('value[x]');
+      onlyRule.types = [{ type: 'Quantity' }];
+      valueX.constrainType(onlyRule, fisher);
+      valueX.sliceIt('type', '$this');
+      // use the exact ElementDefinitionType on value[x] to ensure deep equality
+      const matchingType = valueX.type[0];
+      const valueQuantity = valueX.addSlice('valueQuantity', matchingType);
+      const diff = valueQuantity.calculateDiff();
+      expect(diff.type).toBeDefined();
+      expect(diff.type).toEqual([matchingType]);
     });
   });
